Update dialog/drawer height
